fix: review feedback
BREAKING CHANGE: Renaming allowAnonymous to requireAuthenticatedUser and turning it off by default.  This breaks current MFEs, which will now be required to turn on requireAuthenticatedUser.
